+++ CHANGELOG.md
@@ -4,7 +4,31 @@ All notable changes to this project will be documented in this file,
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.1] - UNRELEASED
+
+### Added
+- Scene Setup commit (15/11/2023)
+- TimeManager and TimeUI added (15/11/2023)
+- Monster transform position from side to side (15/11/2023)
+- Fire bullets logic, bullet script and bullet pool (16/11/2023)
+- Circle, Spin and Pattern Corrutines (17/11/2023)
+- PinkMonsterController and ChestMonsterController, Pink and ChestMonster Prefabs (17/11/2023)
+- LargeBullerPool and LargeBullet Prefab (17/11/2023)
+- Player character and PlayerController (17/11/2023)
+- Bullet Count UI Script and TMP (17/11/2023)
+- AudioSystem (17/11/2023)
+
+### Changed
+
+### Deprecated
+
+### Removed
+- Monster Script (17/11/2023)
+- Time TMP and TimeUI (17/11/2023)
+### Fixed
+- Circle bullet pattern (17/11/2023)
+- ChestMonster bullet pattern (17/11/2023)
+### Security
 
 ## [0.0.0] - 15/11/2023